standardize the rest
