+++ frontend-admin/src/components/clients/modals/index.ts
@@ -1,0 +1,2 @@
+export {CreateModal} from './create';
+export {EditModal} from './edit';
