+++ frontend-admin/src/components/volunteers/modals/index.ts
@@ -1,3 +1,3 @@
-export {NewModal} from './new';
+export {CreateModal} from './create';
 export {EditModal} from './edit';
 export {NotifModal} from './notif';
